Made thing stop breaking under gcc
Both osx vars in the first guard in swap.c were undef so the case would
be true. renameat2 isn't defined for me which is weird since I have a
pretty recent glibc (2.33-5), so I added the ability for syscall to be
used. I also added extra attribute stuff for print_error since
__printflike isn't a thing in gcc.